+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SuboptimalSend` detector: PR [#330](https://github.com/nowarp/misti/pull/330) and PR [#332](https://github.com/nowarp/misti/pull/332)
 - `PreferBinaryReceiver` detector: Issue [#335](https://github.com/nowarp/misti/issues/335)
 - `PreferSenderFunction` detector: Issue [#336](https://github.com/nowarp/misti/issues/336)
+- `ImplicitOpcode` detector: Issue [#338](https://github.com/nowarp/misti/issues/338)
 - Support standalone Tools to run without `CompilationUnit`: PR [#326](https://github.com/nowarp/misti/pull/326)
 - Logger: Detector contexts: PR [#327](https://github.com/nowarp/misti/pull/327)
 - Logger: Print timestamps when `--verbose` is set: Issue [#73](https://github.com/nowarp/misti/issues/73)